Create ECommerceApp
This assignment works with conditional flow to create a simple e-commerce system. It will store and display product information, which a user will then be able to choose from after validating their customer code. It will keep track of all their items in a list and then display the order summary and total at the end.